Generate Theme for `writeExcel` or `writeLatex`

`themeNew` produces themes for `writeExcel` or `writeLatex`.

themeNew(..., default_theme = themeDefaultExcel())

Arguments

...
Named arguments that specify the various output types. See the Theme Arguments section below for a complete listing.

default_theme
a theme object to use for defaults if you're modifying an already existing themes

Theme Arguments

digits

A numeric. How many digits should the data be rounded to? (In Excel, this is excel styling.) Defaults to 0.

digits_numeric

A numeric. How many digits should continuous variable data be rounded to? (In Latex, , this is Latex styling.) Defaults to 2.

digits_final

In Excel, an optional numeric. How many digits should the data be rounded to before being added to Excel?

excel_footer

In Excel, an optional character vector of length 3. The footer text of the file.

excel_freeze_column

In Excel, a numeric. What column should be the last frozen column? Defaults to 1.

excel_header

In Excel, An optional character vector of length 3. The header text of the file.

excel_orientation

In Excel, a character. The orientation of the page if printed. Valid options are: "landscape", and "portrait". Defaults to "landscape".

excel_percent_sign

In Excel, a logical. Should "%" be pasted in each cell that contains a proportion? Defaults to FALSE.

excel_show_grid_lines

In Excel, a logical. Should the default grid lines of the file show? Defaults to FALSE.

excel_table_border

In Excel, an optional list. The formatting of the border around each downbreak. Includes: border_color, and border_style.

font

An optional character. The font to be used.

font_color

In Excel, an optional color. The color of the font.

font_size

An optional numeric. The size of the font.

format_banner_categories

In Excel, a list. How the banner/crossbreak response options should be formatted. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_banner_names

In Excel, an optional list. How the banner/crossbreak variable names should be formatted. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_banner_split

In Excel, an optional list. How should the banner variables be separated? Includes: border_color, border_style

empty_col In Excel, a logical. Should there be an empty column to separate banner variables? Defaults to FALSE.
format_headers

An optional list. How headers should be formatted. If `NULL` headers will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_label_column

In Excel, a list. How the labels column should be formatted. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, wrap_text. In LaTeX, allows you to set col_width only, in inches of the crosstab stub globally.

Includes: col_width A numeric. Width of the label column. Defaults to 40. extend_borders In Excel, a logical. Should the borders created for certain rows extend to the label column? Defaults to FALSE.
format_label_column_exception

In LaTeX, a character vector of columns widths, specified in inches and named after the question alias whose stub they would effect.

format_means

An optional list. How means should be formatted. If `NULL` means will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, name, position_bottom, position_top, valign, and wrap_text.

format_medians

An optional list. How medians should be formatted. If `NULL` medians will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, name, position_bottom, position_top, valign, and wrap_text.

format_min_base

An optional list. If a minimum base size is desired, how variables that fall below that base size should be formatted. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, wrap_text

Includes: mask An optional character to be used to mark cells with base below the min_base. min_base An optional numeric. The minimum acceptable base size for a question.
format_subtitle

An optional list. How the table subtitle should be formatted. If `NULL` the table subtitle will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_subtotals

An optional list. How subtotals should be formatted. If `NULL` subtotals will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_title

An optional list. How the table title should be formatted. If `NULL` the table title will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_totals_column

In Excel, a list. How the totals column should be formatted. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, valign, and wrap_text.

format_totals_row

An optional list. How total rows should be formatted. If `NULL` total rows will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, name, position_bottom, position_top, valign, and wrap_text.

format_var_alias

An optional list. How downbreak variable aliases should be formatted. If `NULL` downbreak variable aliases will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, include_q_number, valign, and wrap_text.

format_var_description

An optional list. How downbreak variable descriptions should be formatted. If `NULL` downbreak variable descriptions will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, include_alias, include_q_number, repeat_for_subs, valign, and wrap_text.

format_var_filtertext

An optional list. How downbreak variable filtertext/notes should be formatted. If `NULL` downbreak variable filtertext/notes will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, include_alias, include_q_number, repeat_for_subs, valign, and wrap_text.

format_var_name

An optional list. How downbreak variable names should be formatted. If `NULL` downbreak variable names will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, include_alias, include_q_number, repeat_for_subs, valign, and wrap_text.

format_var_subname

An optional list. How downbreak subvariable names should be formatted. If `NULL` downbreak subvariable names will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, include_alias, include_q_number, valign, and wrap_text.

format_unweighted_n

An optional list. How unweighted Ns should be formatted. If `NULL` unweighted Ns will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, name, position_bottom, position_fixed, position_top, valign, and wrap_text.

format_weighted_n

An optional list. How weighted Ns should be formatted. If `NULL` weighted Ns will not appear. Includes: background_color, border_bottom, border_color, border_left, border_right, border_style, border_top, decoration, font, font_color, font_size, halign, name, position_bottom, position_fixed, position_top, valign, and wrap_text.

latex_foottext

In Latex, a character. A character string indicating what text should be placed at the top of continuation tables. 'tbc' is a shortcut for 'to be continued.'

latex_headtext

In Latex, a character. A character string indicating what text should be placed at the bottom of continuation tables. 'tbc' is a shortcut for 'to be continued.'

latex_max_lines_for_tabular

In Latex, an integer. What is the maximum number of lines a table can be before it is converted to a longtable? Currently only works on toplines. Defaults to 0.

latex_multirowheaderlines

In Latex, a logical. Should banners allow multi-row headlines? Defaults to FALSE.

latex_round_percentages

In Latex, a logical. In Latex, should percentages be recalculated so they do not exceed 100% where necessary? Defaults to FALSE.

latex_round_percentages_exception

In Latex, an optional character. A list of variable aliases that should have the opposite behaviour of that specified in latex_round_percentages.

latex_table_align

In Latex, a character. A character string indicating what the table alignment should be. Defaults to 'r'.

latex_flip_grids

In Latex, a logical. Categorical arrays will be flipped so that there rows are now transposed to columns.

,
latex_flip_specific_grids

An optional vector of aliases whose presentation should be transposed

logo

An optional list. Information about the logo to be included in the tables.

Includes: file The path to a PNG file that should be used for the logo. Include the extension (.png) for an Excel theme. Exclude the extension for a Latex theme dpi In Excel, a numeric. The image resolution used for conversion between units. Defaults to 300. height In Excel, a numeric. The height of the logo. Defaults to 2. width In Excel, a numeric. The width of the logo. Defaults to 4. startCol In Excel, a numeric. The column coordinate of upper left corner of the logo. Defaults to 1. startRow In Excel, a numeric. The row coordinate of upper left corner of the logo. Defaults to 1. units In Excel, a character. Units of width and height. Valid options are: "cm", "in", and "px." Defaults to "in".
one_per_sheet

A logical. Should each question be on its own sheet/page? Defaults to FALSE.

Subarguments

background_color

In Excel, an optional color. Cell background color.

border_bottom

In Excel, an optional logical. Should there be a border on the bottom?

border_color

In Excel, an optional color. The border color of the relevant cells.

border_left

In Excel, an optional logical. Should there be a border on the left of the relevant cells?

border_right

In Excel, an optional logical. Should there be a border on the right of the relevant cells?

border_style

In Excel, an optional character. The style of the border of the relevant cells. Valid options are: "dashDot", "dashDotDot", "dashed", "dotted", "double", "hair", "medium", "mediumDashDot", "mediumDashDotDot", "mediumDashed", "none", "slantDashDot", "thick", and "thin".

border_top

In Excel, an optional logical. Should there be a border on the top of the relevant cells?

decoration

An optional character vector. Text decorations to be applied to relevant cells. Valid options are: "bold", "italic", "strikeout", "underline", and "underline2".

font

An optional character. The font to be used.

font_color

In Excel, an optional color. The color of the font.

font_size

An optional numeric. The size of the font.

halign

In Excel, an optional character. The horizontal alignment of the text. Valid options are: "center", "left", and "right".

include_alias

A logical. Should the alias of the variable be included with the other information? Defaults to FALSE.

include_q_number

A logical. Should the question number be included with the other information? Defaults to FALSE.

latex_add_parenthesis

In Latex, a logical. Should parenthesis be added surrounding the values? Defaults to FALSE.

latex_adjust

In Latex, an optional character. How should the values be adjusted? Can be missing.

name

A character. The name to be used for the relevant row(s).

position_bottom

In Excel, a logical. Should the relevant row(s) be at the bottom of each table? Defaults to TRUE.

position_fixed

In Excel, a logical. Should the relevant row(s) be fixed at the top of the file with the banner? Defaults to FALSE.

position_top

In Excel, a logical. Should should the relevant row(s) be at the top of each table? Defaults to FALSE.

repeat_for_subs

A logical. Should the information be repeated for each subvariable? Defaults to TRUE.

pagebreak_in_banner

A logical. Allow a banner to be broken midway across pages. Defaults to TRUE. If FALSE, Pushes the page breaking banner sub-table to the next page similar to manually using clearpage

valign

In Excel, an optional character. The vertical alignment of the text. Valid options are: "bottom", "center", and "top".

wrap_text

In Excel, an optional logical. Should the text wrap if it extends beyond the width of the cell? Defaults to TRUE.